+++ jpower-core-server/jpower-system/system-biz/src/main/java/com/wlcb/jpower/service/org/CoreOrgService.java
@@ -38,7 +38,7 @@ public interface CoreOrgService extends BaseService<TbCoreOrg> {
      * @Param [ids]
      * @return java.lang.Integer
      **/
-    Integer listOrgByPids(String ids);
+    long listOrgByPids(String ids);
 
     /**
      * @author 郭丁志
